requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.10
flask-mongoengine==1.0.0
python-dotenv==1.0.1
mongoengine==0.24.2
pymongo==3.12.3
Werkzeug==2.3.7
flask-cors==3.0.10
mongomock==4.1.2
